requirements.txt
Django==2.2.7
django-chunked-upload
